refactor: recentlyVisitedGroup 커스텀 hook으로 변경
초대장에서만 사용하는 기능이기 때문에 parameter를 전달할 필요가 없다고 판단
